only drop db in dev

+++ index.js
@@ -24,9 +24,9 @@ const { generateRegistrationChallenge, parseRegisterRequest } = require("@webaut
 
 const { authenticator } = require("otplib");
 
-app.listen(process.env.PORT != undefined ? process.env.PORT : 80);
+app.listen(process.env.PORT !== undefined ? process.env.PORT : 80);
 console.log("server started");
-db.get("login").drop();
+if (process.env.DEVELOPMENT) db.get("login").drop();
 const webSchema = process.env.WEB_SCHEMA != undefined ? process.env.WEB_SCHEMA : "https";
 
 app.post("/akkount/v1/login", async (req, res) => {
